Create Google Ads MCC Search query report yesterday
manager/MCC account search query report for the previous day. search queries are no longer available for all ad clicks received...so this will not be a full picture of the previous day's search queries. Search queries that are grouped under "Other search terms" are hidden and ad spend and performance from the unknown/hidden/other search terms is not covered/reached by this script.